Added helper methods to create responses and requests
Added a wrapper to the node https.get to hide some of the ugly logic we have to do to get the response body. I considered using axios or requests but since its a small example project I may as well just keep it simple.

Also added a little helper method to add CORS and stringify a given body for API gateway.